+++ packages/kit-plugin-instruction-plan/README.md
@@ -89,100 +89,9 @@ const client = createEmptyClient()
     const transactionPlanResult = await client.sendTransaction(myInstructionPlan);
     ```
 
-## `defaultTransactionPlannerAndExecutorFromRpc` plugin
+## Default Planner and Executor Implementations
 
-This plugin provides default transaction planner and executor implementations using RPC and RPC Subscriptions.
+For ready-to-use transaction planner and executor implementations, see:
 
-### Installation
-
-This plugin requires the `rpc` and `rpcSubscriptions` attributes to be configured on the client.
-
-It also requires a payer to be set on the client or passed as an option to the plugin.
-
-```ts
-import { createEmptyClient } from '@solana/kit';
-import { rpc, generatedPayer, defaultTransactionPlannerAndExecutorFromRpc } from '@solana/kit-plugins';
-
-const client = await createEmptyClient()
-    .use(rpc('https://api.mainnet-beta.solana.com'))
-    .use(generatedPayer())
-    .use(defaultTransactionPlannerAndExecutorFromRpc());
-```
-
-### Options
-
-- `maxConcurrency`: Maximum number of concurrent executions (default: 10).
-- `payer`: Transaction signer for fees (defaults to client's payer if any).
-- `priorityFees`: Priority fees in micro lamports per compute unit.
-- `skipPreflight`: Whether to skip the preflight simulation when sending transactions (default: `false`).
-
-```ts
-const client = await createEmptyClient()
-    .use(rpc('https://api.mainnet-beta.solana.com'))
-    .use(
-        defaultTransactionPlannerAndExecutorFromRpc({
-            maxConcurrency: 5,
-            payer: myPayer,
-            priorityFees: microLamports(1000n),
-        }),
-    );
-```
-
-### Preflight and Compute Unit Estimation
-
-By default, the executor estimates compute units by simulating the transaction before sending it. When estimation is performed, preflight is skipped to avoid a redundant second simulation. When the transaction has an explicit compute unit limit (no estimation needed), preflight runs as the only simulation.
-
-Setting `skipPreflight: true` changes the behavior:
-
-- Preflight is always skipped regardless of whether estimation was performed.
-- If the compute unit estimation simulation fails, the consumed units from the failed simulation are used to set the compute unit limit (with a 10% buffer) so the transaction still reaches the validator. This is useful for debugging failed transactions in an explorer.
-
-| Scenario            | `skipPreflight: false` (default) | `skipPreflight: true`           |
-| ------------------- | -------------------------------- | ------------------------------- |
-| Estimation succeeds | Set CU, skip preflight           | Set CU, skip preflight          |
-| Estimation fails    | Throw                            | Use consumed CU, skip preflight |
-| Explicit CU limit   | Run preflight                    | Skip preflight                  |
-
-### Features
-
-_See the `transactionPlanner` and `transactionPlanExecutor` plugins for available features_.
-
-## `defaultTransactionPlannerAndExecutorFromLitesvm` plugin
-
-This plugin provides default transaction planner and executor implementations using LiteSVM.
-
-### Installation
-
-This plugin requires a LiteSVM instance to be configured on the client.
-
-It also requires a payer to be set on the client or passed as an option to the plugin.
-
-```ts
-import { createEmptyClient } from '@solana/kit';
-import { litesvm, generatedPayer, defaultTransactionPlannerAndExecutorFromLitesvm } from '@solana/kit-plugins';
-
-const client = await createEmptyClient()
-    .use(litesvm())
-    .use(generatedPayer())
-    .use(defaultTransactionPlannerAndExecutorFromLitesvm());
-```
-
-### Options
-
-- `payer`: Transaction signer for fees (defaults to client's payer if any).
-- `priorityFees`: Priority fees in micro lamports per compute unit.
-
-```ts
-const client = await createEmptyClient()
-    .use(litesvm())
-    .use(
-        defaultTransactionPlannerAndExecutorFromLitesvm({
-            payer: myPayer,
-            priorityFees: microLamports(1000n),
-        }),
-    );
-```
-
-### Features
-
-_See the `transactionPlanner` and `transactionPlanExecutor` plugins for available features_.
+- [`@solana/kit-plugin-rpc`](https://www.npmjs.com/package/@solana/kit-plugin-rpc) — provides `rpcTransactionPlanner` and `rpcTransactionPlanExecutor` for RPC-based transaction planning and execution.
+- [`@solana/kit-plugin-litesvm`](https://www.npmjs.com/package/@solana/kit-plugin-litesvm) — provides `litesvmTransactionPlanner` and `litesvmTransactionPlanExecutor` for LiteSVM-based transaction planning and execution.
