+++ packages/kit-plugin-instruction-plan/src/index.ts
@@ -1,3 +1,1 @@
 export * from './core';
-export * from './litesvm';
-export * from './rpc';
