+++ packages/kit-plugin-litesvm/package.json
@@ -48,7 +48,8 @@
         "@solana/kit": "^6.1.0"
     },
     "dependencies": {
-        "@loris-sandbox/litesvm-kit": "^0.5.0"
+        "@loris-sandbox/litesvm-kit": "^0.5.0",
+        "@solana-program/compute-budget": "^0.14.0"
     },
     "license": "MIT",
     "repository": {
