+++ packages/kit-plugin-litesvm/test/transaction-plan-executor.test.ts
@@ -1,3 +1,4 @@
+import type { LiteSVM } from '@loris-sandbox/litesvm-kit';
 import {
     Address,
     createEmptyClient,
@@ -5,46 +6,28 @@ import {
     generateKeyPairSigner,
     setTransactionMessageFeePayerSigner,
     singleInstructionPlan,
-    SingleTransactionPlan,
     singleTransactionPlan,
     SingleTransactionPlanResult,
     SOLANA_ERROR__INSTRUCTION_PLANS__FAILED_TO_EXECUTE_TRANSACTION_PLAN,
     SOLANA_ERROR__TRANSACTION__FAILED_WHEN_SIMULATING_TO_ESTIMATE_COMPUTE_LIMIT,
     SolanaError,
-    TransactionSigner,
 } from '@solana/kit';
-import type { LiteSVM } from '@solana/kit-plugin-litesvm';
 import { describe, expect, it, vi } from 'vitest';
 
-import { defaultTransactionPlannerAndExecutorFromLitesvm } from '../src';
+import { litesvmTransactionPlanExecutor, litesvmTransactionPlanner } from '../src';
 
 const MOCK_BLOCKHASH = { blockhash: '11111111111111111111111111111111', lastValidBlockHeight: 0n };
 const MOCK_INSTRUCTION = {
     programAddress: '11111111111111111111111111111111' as Address,
 };
 
-describe('defaultTransactionPlannerAndExecutorFromLitesvm', () => {
-    it('provides a default transactionPlanner and transactionPlanExecutor on the client', () => {
-        const payer = {} as TransactionSigner;
+describe('litesvmTransactionPlanExecutor', () => {
+    it('provides a transactionPlanExecutor on the client', () => {
         const svm = {} as LiteSVM;
         const client = createEmptyClient()
-            .use(() => ({ payer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm());
-        expect(client).toHaveProperty('transactionPlanner');
+            .use(() => ({ svm }))
+            .use(litesvmTransactionPlanExecutor());
         expect(client).toHaveProperty('transactionPlanExecutor');
-    });
-
-    it('uses the provided payer as fee payer when planning transactions', async () => {
-        const payer = await generateKeyPairSigner();
-        const svm = {} as LiteSVM;
-        const client = createEmptyClient()
-            .use(() => ({ payer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm());
-
-        const instructionPlan = singleInstructionPlan(MOCK_INSTRUCTION);
-        const transactionPlan = (await client.transactionPlanner(instructionPlan)) as SingleTransactionPlan;
-        expect(transactionPlan.kind).toBe('single');
-        expect(transactionPlan.message.feePayer).toBe(payer);
     });
 
     it('uses the SVM instance to send transactions', async () => {
@@ -54,7 +37,8 @@ describe('defaultTransactionPlannerAndExecutorFromLitesvm', () => {
         const svm = { latestBlockhashLifetime, sendTransaction } as unknown as LiteSVM;
         const client = createEmptyClient()
             .use(() => ({ payer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm());
+            .use(litesvmTransactionPlanner())
+            .use(litesvmTransactionPlanExecutor());
 
         const instructionPlan = singleInstructionPlan(MOCK_INSTRUCTION);
         const transactionPlan = await client.transactionPlanner(instructionPlan);
@@ -79,7 +63,7 @@ describe('defaultTransactionPlannerAndExecutorFromLitesvm', () => {
         const svm = { latestBlockhashLifetime, sendTransaction } as unknown as LiteSVM;
         const client = createEmptyClient()
             .use(() => ({ payer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm());
+            .use(litesvmTransactionPlanExecutor());
 
         const transactionPlan = singleTransactionPlan(
             setTransactionMessageFeePayerSigner(payer, createTransactionMessage({ version: 0 })),
@@ -101,7 +85,7 @@ describe('defaultTransactionPlannerAndExecutorFromLitesvm', () => {
         const svm = { latestBlockhashLifetime, sendTransaction } as unknown as LiteSVM;
         const client = createEmptyClient()
             .use(() => ({ payer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm());
+            .use(litesvmTransactionPlanExecutor());
 
         const transactionPlan = singleTransactionPlan(
             setTransactionMessageFeePayerSigner(payer, createTransactionMessage({ version: 0 })),
@@ -115,38 +99,6 @@ describe('defaultTransactionPlannerAndExecutorFromLitesvm', () => {
 
     it('requires an svm instance on the client', () => {
         // @ts-expect-error Missing svm instance on the client.
-        expect(() => createEmptyClient().use(defaultTransactionPlannerAndExecutorFromLitesvm())).toThrow();
-    });
-
-    it('requires a payer on the client by default', () => {
-        const svm = {} as LiteSVM;
-        expect(() =>
-            createEmptyClient()
-                .use(() => ({ svm }))
-                .use(defaultTransactionPlannerAndExecutorFromLitesvm()),
-        ).toThrow();
-    });
-
-    it('also accepts a payer directly', () => {
-        const payer = {} as TransactionSigner;
-        const svm = {} as LiteSVM;
-        expect(() =>
-            createEmptyClient()
-                .use(() => ({ svm }))
-                .use(defaultTransactionPlannerAndExecutorFromLitesvm({ payer })),
-        ).not.toThrow();
-    });
-
-    it('uses the provided payer over the one set on the client', async () => {
-        const [clientPayer, explicitPayer] = await Promise.all([generateKeyPairSigner(), generateKeyPairSigner()]);
-        const svm = {} as LiteSVM;
-        const client = createEmptyClient()
-            .use(() => ({ payer: clientPayer, svm }))
-            .use(defaultTransactionPlannerAndExecutorFromLitesvm({ payer: explicitPayer }));
-
-        const instructionPlan = singleInstructionPlan(MOCK_INSTRUCTION);
-        const transactionPlan = (await client.transactionPlanner(instructionPlan)) as SingleTransactionPlan;
-        expect(transactionPlan.kind).toBe('single');
-        expect(transactionPlan.message.feePayer).toBe(explicitPayer);
+        expect(() => createEmptyClient().use(litesvmTransactionPlanExecutor())).toThrow();
     });
 });
