+++ packages/kit-plugin-rpc/README.md
@@ -80,3 +80,82 @@ const client = createEmptyClient().use(localhostRpc());
 ### Features
 
 _See the `rpc` plugin for available features_.
+
+## `rpcTransactionPlanner` plugin
+
+This plugin provides a default transaction planner that creates transaction messages with a fee payer, a provisory compute unit limit, and optional priority fees.
+
+### Installation
+
+This plugin requires a payer to be set on the client or passed as an option.
+
+```ts
+import { createEmptyClient } from '@solana/kit';
+import { rpc, rpcTransactionPlanner, rpcTransactionPlanExecutor } from '@solana/kit-plugin-rpc';
+import { generatedPayer } from '@solana/kit-plugin-payer';
+
+const client = await createEmptyClient()
+    .use(rpc('https://api.mainnet-beta.solana.com'))
+    .use(generatedPayer())
+    .use(rpcTransactionPlanner())
+    .use(rpcTransactionPlanExecutor());
+```
+
+### Options
+
+- `payer`: Transaction signer for fees (defaults to client's payer if any).
+- `priorityFees`: Priority fees in micro lamports per compute unit.
+
+### Features
+
+- `transactionPlanner`: A function that plans instructions into transaction messages.
+    ```ts
+    const transactionPlan = await client.transactionPlanner(myInstructionPlan);
+    ```
+
+## `rpcTransactionPlanExecutor` plugin
+
+This plugin provides a default transaction plan executor that estimates compute units, signs, and sends transactions via RPC.
+
+### Installation
+
+This plugin requires `rpc` and `rpcSubscriptions` to be configured on the client.
+
+```ts
+import { createEmptyClient } from '@solana/kit';
+import { rpc, rpcTransactionPlanner, rpcTransactionPlanExecutor } from '@solana/kit-plugin-rpc';
+import { generatedPayer } from '@solana/kit-plugin-payer';
+
+const client = await createEmptyClient()
+    .use(rpc('https://api.mainnet-beta.solana.com'))
+    .use(generatedPayer())
+    .use(rpcTransactionPlanner())
+    .use(rpcTransactionPlanExecutor());
+```
+
+### Options
+
+- `maxConcurrency`: Maximum number of concurrent executions (default: 10).
+- `skipPreflight`: Whether to skip the preflight simulation when sending transactions (default: `false`).
+
+### Features
+
+- `transactionPlanExecutor`: A function that executes planned transactions.
+    ```ts
+    const transactionPlanResult = await client.transactionPlanExecutor(myTransactionPlan);
+    ```
+
+### Preflight and Compute Unit Estimation
+
+By default, the executor estimates compute units by simulating the transaction before sending it. When estimation is performed, preflight is skipped to avoid a redundant second simulation. When the transaction has an explicit compute unit limit (no estimation needed), preflight runs as the only simulation.
+
+Setting `skipPreflight: true` changes the behavior:
+
+- Preflight is always skipped regardless of whether estimation was performed.
+- If the compute unit estimation simulation fails, the consumed units from the failed simulation are used to set the compute unit limit (with a 10% buffer) so the transaction still reaches the validator. This is useful for debugging failed transactions in an explorer.
+
+| Scenario            | `skipPreflight: false` (default) | `skipPreflight: true`           |
+| ------------------- | -------------------------------- | ------------------------------- |
+| Estimation succeeds | Set CU, skip preflight           | Set CU, skip preflight          |
+| Estimation fails    | Throw                            | Use consumed CU, skip preflight |
+| Explicit CU limit   | Run preflight                    | Skip preflight                  |
