+++ packages/kit-plugin-rpc/package.json
@@ -47,6 +47,9 @@
     "peerDependencies": {
         "@solana/kit": "^6.1.0"
     },
+    "dependencies": {
+        "@solana-program/compute-budget": "^0.14.0"
+    },
     "license": "MIT",
     "repository": {
         "type": "git",
